move gitHubTests from resolvers to models
